make constants dir + move button groups there
