+++ src/constants/index.js
@@ -1,0 +1,5 @@
+export {
+    makeDeveloperButtons,
+    makeEditorButtons,
+    makeEditorialButtons
+} from './toolbarButtonGroups'
